[Android] Moved DoesDefaultSearchEngineHaveLogo overriding
Chromium change:
https://chromium.googlesource.com/chromium/src/+/8102e6c428f77caaeeeac173528343f500106424

Fix a flaky test about non-Google search engine

The test LocationBarTest#testOmniboxSearchEngineLogo_unfocusedOnSRP_\
nonGoogleSearchEngine frequently times out while loading a logo image
to be displayed on NTP. It often takes longer, and keeps the test from
moving forward. Since the NTP logo itself is not relevant to the test,
this CL replaces the image with Google logo which can be loaded in
a predictable way to let the test go as intended and pass.

More specifically, TemplateUrlServiceFactory#
doesDefaultSearchEngineHaveLogo is moved to TemplateUrlService and
mocked for the test. This is used by NTP not by Omnibox so won't
affect the purpose of the test.

Bug: 1227573